python-backend/requirements.txt
fastapi
uvicorn[standard]

opencv-python
numpy
Pillow

insightface
gfpgan
facexlib
onnxruntime

supabase
python-dotenv
httpx

python-multipart
aiofiles

loguru

tqdm
